+++ cmd/gitserver/internal/git/BUILD.bazel
@@ -7,6 +7,7 @@ go_library(
         "cleanup.go",
         "config.go",
         "git.go",
+        "object.go",
         "type.go",
     ],
     importpath = "github.com/sourcegraph/sourcegraph/cmd/gitserver/internal/git",
@@ -15,19 +16,32 @@ go_library(
         "//cmd/gitserver/internal/common",
         "//cmd/gitserver/internal/executil",
         "//cmd/gitserver/internal/gitserverfs",
+        "//internal/api",
         "//internal/conf",
         "//internal/fileutil",
         "//internal/gitserver/gitdomain",
         "//internal/syncx",
+        "//internal/trace",
         "//internal/wrexec",
         "//lib/errors",
         "@com_github_sourcegraph_log//:log",
+        "@io_opentelemetry_go_otel//attribute",
     ],
 )
 
 go_test(
     name = "git_test",
-    srcs = ["git_test.go"],
+    srcs = [
+        "git_test.go",
+        "object_test.go",
+    ],
     embed = [":git"],
-    deps = ["//cmd/gitserver/internal/common"],
+    deps = [
+        "//cmd/gitserver/internal/common",
+        "//internal/api",
+        "//internal/gitserver/gitdomain",
+        "//internal/wrexec",
+        "//lib/errors",
+        "@com_github_google_go_cmp//cmp",
+    ],
 )
