+++ cmd/gitserver/internal/server.go
@@ -7,7 +7,6 @@ import (
 	"container/list"
 	"context"
 	"encoding/gob"
-	"encoding/hex"
 	"encoding/json"
 	"fmt"
 	"io"
@@ -689,7 +688,7 @@ func (s *Server) handleArchive(w http.ResponseWriter, r *http.Request) {
 		log.Strings("path", pathspecs),
 	)
 
-	if err := checkSpecArgSafety(treeish); err != nil {
+	if err := git.CheckSpecArgSafety(treeish); err != nil {
 		w.WriteHeader(http.StatusBadRequest)
 		s.Logger.Error("gitserver.archive.CheckSpecArgSafety", log.Error(err))
 		return
@@ -2263,7 +2262,7 @@ func (s *Server) handleGetObject(w http.ResponseWriter, r *http.Request) {
 	// Log which actor is accessing the repo.
 	accesslog.Record(r.Context(), string(req.Repo), log.String("objectname", req.ObjectName))
 
-	obj, err := getObject(r.Context(), s.RecordingCommandFactory, s.ReposDir, getObjectType, revParse, req.Repo, req.ObjectName)
+	obj, err := git.GetObject(r.Context(), s.RecordingCommandFactory, s.ReposDir, req.Repo, req.ObjectName)
 	if err != nil {
 		http.Error(w, errors.Wrap(err, "getting object").Error(), http.StatusInternalServerError)
 		return
@@ -2277,95 +2276,4 @@ func (s *Server) handleGetObject(w http.ResponseWriter, r *http.Request) {
 		http.Error(w, err.Error(), http.StatusInternalServerError)
 		return
 	}
-}
-
-func getObject(ctx context.Context, rcf *wrexec.RecordingCommandFactory, reposDir string, getObjectType getObjectTypeFunc, revParse revParseFunc, repo api.RepoName, objectName string) (_ *gitdomain.GitObject, err error) {
-	tr, ctx := trace.New(ctx, "GetObject",
-		attribute.String("objectName", objectName))
-	defer tr.EndWithErr(&err)
-
-	if err := checkSpecArgSafety(objectName); err != nil {
-		return nil, err
-	}
-
-	dir := gitserverfs.RepoDirFromName(reposDir, repo)
-
-	sha, err := revParse(ctx, rcf, repo, dir, objectName)
-	if err != nil {
-		if gitdomain.IsRepoNotExist(err) {
-			return nil, err
-		}
-		if strings.Contains(sha, "unknown revision") {
-			return nil, &gitdomain.RevisionNotFoundError{Repo: repo, Spec: objectName}
-		}
-		return nil, err
-	}
-
-	sha = strings.TrimSpace(sha)
-	if !gitdomain.IsAbsoluteRevision(sha) {
-		if sha == "HEAD" {
-			// We don't verify the existence of HEAD, but if HEAD doesn't point to anything
-			// git just returns `HEAD` as the output of rev-parse. An example where this
-			// occurs is an empty repository.
-			return nil, &gitdomain.RevisionNotFoundError{Repo: repo, Spec: objectName}
-		}
-		return nil, &gitdomain.BadCommitError{Spec: objectName, Commit: api.CommitID(sha), Repo: repo}
-	}
-
-	oid, err := decodeOID(sha)
-	if err != nil {
-		return nil, errors.Wrap(err, "decoding oid")
-	}
-
-	objectType, err := getObjectType(ctx, rcf, repo, dir, oid.String())
-	if err != nil {
-		return nil, errors.Wrap(err, "getting object type")
-	}
-
-	return &gitdomain.GitObject{
-		ID:   oid,
-		Type: objectType,
-	}, nil
-
-}
-
-type getObjectTypeFunc func(ctx context.Context, rcf *wrexec.RecordingCommandFactory, repo api.RepoName, dir common.GitDir, objectID string) (gitdomain.ObjectType, error)
-
-// getObjectType returns the object type given an objectID.
-func getObjectType(ctx context.Context, rcf *wrexec.RecordingCommandFactory, repo api.RepoName, dir common.GitDir, objectID string) (gitdomain.ObjectType, error) {
-	cmd := exec.Command("git", "cat-file", "-t", "--", objectID)
-	dir.Set(cmd)
-	wrappedCmd := rcf.WrapWithRepoName(context.Background(), log.NoOp(), repo, cmd)
-	out, err := wrappedCmd.CombinedOutput()
-	if err != nil {
-		return "", errors.WithMessage(err, fmt.Sprintf("git command %v failed (output: %q)", wrappedCmd.Args, out))
-	}
-
-	objectType := gitdomain.ObjectType(bytes.TrimSpace(out))
-	return objectType, nil
-}
-
-type revParseFunc func(ctx context.Context, rcf *wrexec.RecordingCommandFactory, repo api.RepoName, dir common.GitDir, rev string) (string, error)
-
-// revParse will run rev-parse on the given rev.
-func revParse(ctx context.Context, rcf *wrexec.RecordingCommandFactory, repo api.RepoName, dir common.GitDir, rev string) (string, error) {
-	cmd := exec.Command("git", "rev-parse", rev)
-	dir.Set(cmd)
-	wrappedCmd := rcf.WrapWithRepoName(context.Background(), log.NoOp(), repo, cmd)
-	out, err := wrappedCmd.CombinedOutput()
-	if err != nil {
-		return "", errors.WithMessage(err, fmt.Sprintf("git command %v failed (output: %q)", wrappedCmd.Args, out))
-	}
-
-	return string(out), nil
-}
-
-func decodeOID(sha string) (gitdomain.OID, error) {
-	oidBytes, err := hex.DecodeString(sha)
-	if err != nil {
-		return gitdomain.OID{}, err
-	}
-	var oid gitdomain.OID
-	copy(oid[:], oidBytes)
-	return oid, nil
 }
